requirements.txt
flask-restful
sqlalchemy
flask_sqlalchemy
flask_migrate
python-dotenv
Flask-Migrate
sqlalchemy-utils